Fix ci
