+++ packages/twenty-server/src/engine/core-modules/file/file-upload/services/file-upload.service.ts
@@ -7,7 +7,7 @@ import { pipeline } from 'stream/promises';
 import { msg } from '@lingui/core/macro';
 import { isNonEmptyString } from '@sniptt/guards';
 import bytes from 'bytes';
-import { ApiPath, FileFolder } from 'twenty-shared/types';
+import { FileFolder } from 'twenty-shared/types';
 import { isDefined } from 'twenty-shared/utils';
 import { Repository } from 'typeorm';
 import { v4 } from 'uuid';
@@ -15,33 +15,23 @@ import { v4 } from 'uuid';
 import { settings } from 'src/engine/constants/settings';
 import { ApplicationEntity } from 'src/engine/core-modules/application/application.entity';
 import { ApplicationService } from 'src/engine/core-modules/application/application.service';
-import { FileUploadTokenJwtPayload } from 'src/engine/core-modules/auth/types/file-upload-token-jwt-payload.type';
-import { JwtTokenTypeEnum } from 'src/engine/core-modules/auth/types/jwt-token-type.enum';
 import { FileStorageService } from 'src/engine/core-modules/file-storage/services/file-storage.service';
-import { FileDTO } from 'src/engine/core-modules/file/dtos/file.dto';
 import { FileWithSignedUrlDTO } from 'src/engine/core-modules/file/dtos/file-with-sign-url.dto';
 import { FileEntity } from 'src/engine/core-modules/file/entities/file.entity';
-import { FILE_CONTENT_SNIFF_BYTE_COUNT } from 'src/engine/core-modules/file/file-upload/constants/file-content-sniff.constant';
 import { FileUploadTargetDTO } from 'src/engine/core-modules/file/file-upload/dtos/file-upload-target.dto';
 import {
   FileUploadException,
   FileUploadExceptionCode,
 } from 'src/engine/core-modules/file/file-upload/file-upload.exception';
+import { FileUploadCompletionService } from 'src/engine/core-modules/file/file-upload/services/file-upload-completion.service';
+import { FileUploadTargetService } from 'src/engine/core-modules/file/file-upload/services/file-upload-target.service';
 import { FileUrlService } from 'src/engine/core-modules/file/file-url/file-url.service';
-import { FileSettings } from 'src/engine/core-modules/file/types/file-settings.types';
 import { FILE_STATUS } from 'src/engine/core-modules/file/types/file-status.types';
 import { buildFileInfo } from 'src/engine/core-modules/file/utils/build-file-info.utils';
-import { extractFileInfoOrThrow } from 'src/engine/core-modules/file/utils/extract-file-info-or-throw.utils';
-import { fileFolderConfigs } from 'src/engine/core-modules/file/interfaces/file-folder.interface';
 import { removeFileFolderFromFileEntityPath } from 'src/engine/core-modules/file/utils/remove-file-folder-from-file-entity-path.utils';
-import { sanitizeFile } from 'src/engine/core-modules/file/utils/sanitize-file.utils';
-import { JwtWrapperService } from 'src/engine/core-modules/jwt/services/jwt-wrapper.service';
-import { TwentyConfigService } from 'src/engine/core-modules/twenty-config/twenty-config.service';
 import { FieldMetadataEntity } from 'src/engine/metadata-modules/field-metadata/field-metadata.entity';
 import { InjectWorkspaceScopedRepository } from 'src/engine/twenty-orm/workspace-scoped-repository/inject-workspace-scoped-repository.decorator';
 import { WorkspaceScopedRepository } from 'src/engine/twenty-orm/workspace-scoped-repository/workspace-scoped-repository';
-import { readReadablePrefix } from 'src/utils/read-readable-prefix';
-import { streamToBuffer } from 'src/utils/stream-to-buffer';
 
 export const DIRECT_UPLOAD_FILE_FOLDERS = [
   FileFolder.FilesField,
@@ -51,38 +41,13 @@ export const DIRECT_UPLOAD_FILE_FOLDERS = [
   FileFolder.EmailImage,
 ] as const;
 
-const DIRECT_UPLOAD_CONTENT_TYPE = 'application/octet-stream';
-
-export type BatchUploadTargetRequest = {
-  workspaceId: string;
-  applicationUniversalIdentifier: string;
-  applicationId?: string;
-  fileFolder: FileFolder;
-  resourcePath: string;
-  size: number;
-  settings: FileSettings;
-};
-
-export type BatchCompleteUploadRequest = {
-  workspaceId: string;
-  applicationUniversalIdentifier: string;
-  file: FileEntity;
-};
-
-export type BatchFileResult<TValue> =
-  | { success: true; value: TValue }
-  | { success: false; error: string };
-
-const toBatchErrorMessage = (error: unknown): string =>
-  error instanceof Error ? error.message : String(error);
-
 @Injectable()
 export class FileUploadService {
   constructor(
     private readonly fileStorageService: FileStorageService,
     private readonly fileUrlService: FileUrlService,
-    private readonly jwtWrapperService: JwtWrapperService,
-    private readonly twentyConfigService: TwentyConfigService,
+    private readonly fileUploadTargetService: FileUploadTargetService,
+    private readonly fileUploadCompletionService: FileUploadCompletionService,
     private readonly applicationService: ApplicationService,
     @InjectRepository(ApplicationEntity)
     private readonly applicationRepository: Repository<ApplicationEntity>,
@@ -162,7 +127,7 @@ export class FileUploadService {
       },
     });
 
-    return this.buildUploadTarget({
+    return this.fileUploadTargetService.buildUploadTarget({
       workspaceId,
       fileId,
       fileFolder,
@@ -173,230 +138,6 @@ export class FileUploadService {
     });
   }
 
-  private async buildUploadTarget({
-    workspaceId,
-    fileId,
-    fileFolder,
-    applicationUniversalIdentifier,
-    resourcePath,
-    contentType,
-    size,
-  }: {
-    workspaceId: string;
-    fileId: string;
-    fileFolder: FileFolder;
-    applicationUniversalIdentifier: string;
-    resourcePath: string;
-    contentType: string;
-    size: number;
-  }): Promise<FileUploadTargetDTO> {
-    const expiresInSeconds = this.twentyConfigService.get(
-      'STORAGE_S3_PRESIGNED_URL_EXPIRES_IN',
-    );
-    const expiresAt = new Date(Date.now() + expiresInSeconds * 1000);
-
-    const presignedUploadUrl =
-      await this.fileStorageService.getPresignedUploadUrl({
-        fileFolder,
-        applicationUniversalIdentifier,
-        workspaceId,
-        resourcePath,
-        contentType,
-        contentLength: size,
-        expiresInSeconds,
-      });
-
-    if (isDefined(presignedUploadUrl)) {
-      return {
-        fileId,
-        uploadUrl: presignedUploadUrl,
-        contentType,
-        expiresAt,
-      };
-    }
-
-    // No presign support (local storage, or S3 without presign enabled):
-    // fall back to the token-authenticated streaming endpoint on the server.
-    const payload: FileUploadTokenJwtPayload = {
-      workspaceId,
-      fileId,
-      sub: workspaceId,
-      type: JwtTokenTypeEnum.FILE_UPLOAD,
-    };
-
-    const token = await this.jwtWrapperService.signAsyncOrThrow(payload, {
-      expiresIn: expiresInSeconds,
-    });
-
-    const serverUrl = this.twentyConfigService.get('SERVER_URL');
-
-    return {
-      fileId,
-      uploadUrl: `${serverUrl}/${ApiPath.FileUpload}/${fileId}?token=${token}`,
-      // octet-stream keeps the request body away from the server's json/text
-      // body parsers; the real mime type is already on the file record.
-      contentType: 'application/octet-stream',
-      expiresAt,
-    };
-  }
-
-  async createUploadTargetsBatch(
-    requests: BatchUploadTargetRequest[],
-  ): Promise<BatchFileResult<FileUploadTargetDTO>[]> {
-    return Promise.all(
-      requests.map(async (request) => {
-        try {
-          const pendingFile = await this.fileStorageService.createPendingFile({
-            fileFolder: request.fileFolder,
-            applicationUniversalIdentifier:
-              request.applicationUniversalIdentifier,
-            applicationId: request.applicationId,
-            workspaceId: request.workspaceId,
-            resourcePath: request.resourcePath,
-            fileId: v4(),
-            size: request.size,
-            mimeType: DIRECT_UPLOAD_CONTENT_TYPE,
-            settings: request.settings,
-          });
-
-          const value = await this.buildUploadTarget({
-            workspaceId: request.workspaceId,
-            fileId: pendingFile.id,
-            fileFolder: request.fileFolder,
-            applicationUniversalIdentifier:
-              request.applicationUniversalIdentifier,
-            resourcePath: request.resourcePath,
-            contentType: DIRECT_UPLOAD_CONTENT_TYPE,
-            size: request.size,
-          });
-
-          return { success: true as const, value };
-        } catch (error) {
-          return { success: false as const, error: toBatchErrorMessage(error) };
-        }
-      }),
-    );
-  }
-
-  async completeUploadsBatch(
-    requests: BatchCompleteUploadRequest[],
-  ): Promise<BatchFileResult<FileDTO>[]> {
-    return Promise.all(
-      requests.map(async (request) => {
-        try {
-          const value = await this.finalizeUploadedFile(request);
-
-          return { success: true as const, value };
-        } catch (error) {
-          return { success: false as const, error: toBatchErrorMessage(error) };
-        }
-      }),
-    );
-  }
-
-  private async finalizeUploadedFile({
-    workspaceId,
-    applicationUniversalIdentifier,
-    file,
-  }: BatchCompleteUploadRequest): Promise<FileDTO> {
-    const [fileFolder] = file.path.split('/');
-    const resourcePath = removeFileFolderFromFileEntityPath(file.path);
-
-    const storageLocation = {
-      fileFolder: fileFolder as FileFolder,
-      applicationUniversalIdentifier,
-      workspaceId,
-      resourcePath,
-    };
-
-    const metadata =
-      await this.fileStorageService.getFileMetadata(storageLocation);
-
-    if (!isDefined(metadata)) {
-      throw new FileUploadException(
-        `File "${file.path}" has not been uploaded to storage yet.`,
-        FileUploadExceptionCode.FILE_NOT_UPLOADED,
-        {
-          userFriendlyMessage: msg`The file has not been uploaded yet. Please upload it before confirming.`,
-        },
-      );
-    }
-
-    const declaredSize = Number(file.size);
-
-    if (metadata.size !== declaredSize) {
-      throw new FileUploadException(
-        `File "${file.path}" has ${metadata.size} bytes in storage but ${declaredSize} were declared.`,
-        FileUploadExceptionCode.FILE_SIZE_MISMATCH,
-        {
-          userFriendlyMessage: msg`The uploaded file does not match the declared size. Please retry the upload.`,
-        },
-      );
-    }
-
-    const mimeType = await this.detectUploadedMimeTypeOrThrow({
-      ...storageLocation,
-      filename: file.path,
-    });
-
-    this.assertMimeTypeAllowedForFolder(fileFolder as FileFolder, mimeType);
-
-    const size = await this.sanitizeUploadedFileIfNeeded({
-      storageLocation,
-      mimeType,
-      size: metadata.size,
-    });
-
-    await this.fileRepository.update(
-      workspaceId,
-      { id: file.id },
-      { status: FILE_STATUS.UPLOADED, mimeType, size },
-    );
-
-    return { id: file.id, path: file.path, size, createdAt: file.createdAt };
-  }
-
-  private async sanitizeUploadedFileIfNeeded({
-    storageLocation,
-    mimeType,
-    size,
-  }: {
-    storageLocation: {
-      fileFolder: FileFolder;
-      applicationUniversalIdentifier: string;
-      workspaceId: string;
-      resourcePath: string;
-    };
-    mimeType: string;
-    size: number;
-  }): Promise<number> {
-    if (mimeType !== 'image/svg+xml') {
-      return size;
-    }
-
-    const stream = await this.fileStorageService.readFile(storageLocation);
-    const sanitizedFile = sanitizeFile({
-      file: await streamToBuffer(stream),
-      ext: 'svg',
-      mimeType,
-    });
-
-    const sanitizedBuffer = Buffer.isBuffer(sanitizedFile)
-      ? sanitizedFile
-      : Buffer.from(sanitizedFile);
-
-    await this.fileStorageService.writeFileStream({
-      ...storageLocation,
-      stream: Readable.from(sanitizedBuffer),
-      mimeType,
-    });
-
-    return sanitizedBuffer.length;
-  }
-
-  // Streams the request body straight to the storage driver, bounded by the
-  // size declared at createFileUpload time. Memory usage stays constant no
-  // matter how large the file is.
   async receiveFileStream({
     workspaceId,
     fileId,
@@ -459,8 +200,6 @@ export class FileUploadService {
         }),
       ]);
     } catch (error) {
-      // The storage-side pipeline can lose the rejection race to the limiter:
-      // surface the size violation over the resulting stream teardown error.
       if (receivedBytes > declaredSize) {
         throw new FileUploadException(
           `Upload exceeds declared size of ${declaredSize} bytes`,
@@ -475,10 +214,6 @@ export class FileUploadService {
     }
 
     if (receivedBytes !== declaredSize) {
-      // The short object stays in storage but the record stays PENDING, so it
-      // can never be served or attached: the client retries against the same
-      // upload url (overwriting it), and the pending-file cleanup cron
-      // (follow-up PR) reaps whatever is abandoned.
       throw new FileUploadException(
         `Uploaded ${receivedBytes} bytes but ${declaredSize} were declared`,
         FileUploadExceptionCode.FILE_SIZE_MISMATCH,
@@ -489,8 +224,6 @@ export class FileUploadService {
     }
   }
 
-  // Verifies the bytes actually landed in storage with the declared size and
-  // flips the file to UPLOADED. Idempotent: confirming twice is a no-op.
   async completeFileUpload({
     workspaceId,
     fileId,
@@ -501,8 +234,6 @@ export class FileUploadService {
     const file = await this.findFileOrThrow({ workspaceId, fileId });
     const [fileFolder] = file.path.split('/');
 
-    // Restrict to files created through createFileUpload so this mutation
-    // cannot be used to mint signed download urls for arbitrary files.
     if (
       !DIRECT_UPLOAD_FILE_FOLDERS.includes(
         fileFolder as (typeof DIRECT_UPLOAD_FILE_FOLDERS)[number],
@@ -518,9 +249,6 @@ export class FileUploadService {
     }
 
     if (file.status === FILE_STATUS.UPLOADED) {
-      // Idempotent retry of a confirm that already succeeded. Only files not
-      // yet attached to a record qualify: this cannot be used to mint signed
-      // urls for files that went through the legacy flow and got attached.
       if (!file.settings?.isTemporaryFile) {
         throw new FileUploadException(
           `File ${fileId} is not awaiting an upload confirmation`,
@@ -543,106 +271,23 @@ export class FileUploadService {
       file,
     });
 
-    const metadata = await this.fileStorageService.getFileMetadata({
-      fileFolder: fileFolder as FileFolder,
-      applicationUniversalIdentifier: application.universalIdentifier,
-      workspaceId,
-      resourcePath,
-    });
-
-    if (!isDefined(metadata)) {
-      throw new FileUploadException(
-        `File ${fileId} has not been uploaded to storage yet`,
-        FileUploadExceptionCode.FILE_NOT_UPLOADED,
-        {
-          userFriendlyMessage: msg`The file has not been uploaded yet. Please upload it before confirming.`,
+    const completedFile =
+      await this.fileUploadCompletionService.completeUploadedFile({
+        workspaceId,
+        file,
+        storageLocation: {
+          fileFolder: fileFolder as FileFolder,
+          applicationUniversalIdentifier: application.universalIdentifier,
+          workspaceId,
+          resourcePath,
         },
-      );
-    }
-
-    if (metadata.size !== Number(file.size)) {
-      throw new FileUploadException(
-        `File ${fileId} has ${metadata.size} bytes in storage but ${file.size} were declared`,
-        FileUploadExceptionCode.FILE_SIZE_MISMATCH,
-        {
-          userFriendlyMessage: msg`The uploaded file does not match the declared size. Please retry the upload.`,
-        },
-      );
-    }
-
-    const mimeType = await this.detectUploadedMimeTypeOrThrow({
-      fileFolder: fileFolder as FileFolder,
-      applicationUniversalIdentifier: application.universalIdentifier,
-      workspaceId,
-      resourcePath,
-      filename: file.path,
-    });
-
-    this.assertMimeTypeAllowedForFolder(fileFolder as FileFolder, mimeType);
-
-    await this.fileRepository.update(
-      workspaceId,
-      { id: fileId },
-      { status: FILE_STATUS.UPLOADED, mimeType },
-    );
+      });
 
     return this.toFileWithSignedUrl({
-      file: { ...file, status: FILE_STATUS.UPLOADED, mimeType },
+      file: { ...file, ...completedFile, status: FILE_STATUS.UPLOADED },
       fileFolder: fileFolder as FileFolder,
       workspaceId,
     });
-  }
-
-  private async detectUploadedMimeTypeOrThrow({
-    fileFolder,
-    applicationUniversalIdentifier,
-    workspaceId,
-    resourcePath,
-    filename,
-  }: {
-    fileFolder: FileFolder;
-    applicationUniversalIdentifier: string;
-    workspaceId: string;
-    resourcePath: string;
-    filename: string;
-  }): Promise<string> {
-    const stream = await this.fileStorageService.readFile({
-      fileFolder,
-      applicationUniversalIdentifier,
-      workspaceId,
-      resourcePath,
-    });
-
-    const prefix = await readReadablePrefix(
-      stream,
-      FILE_CONTENT_SNIFF_BYTE_COUNT,
-    );
-
-    const { mimeType } = await extractFileInfoOrThrow({
-      file: prefix,
-      filename,
-    });
-
-    return mimeType;
-  }
-
-  private assertMimeTypeAllowedForFolder(
-    fileFolder: FileFolder,
-    mimeType: string,
-  ): void {
-    const { allowedMimeTypes } = fileFolderConfigs[fileFolder];
-
-    if (!allowedMimeTypes || allowedMimeTypes.includes(mimeType)) {
-      return;
-    }
-
-    throw new FileUploadException(
-      `MIME type ${mimeType} is not allowed in file folder ${fileFolder}`,
-      FileUploadExceptionCode.BAD_REQUEST,
-      {
-        userFriendlyMessage: msg`This file format is not supported.`,
-      },
-    );
   }
 
   private async resolveUploadLocation({
