+++ packages/twenty-server/src/engine/core-modules/file/file-upload/types/batch-file-result.type.ts
@@ -1,0 +1,3 @@
+export type BatchFileResult<TValue> =
+  | { success: true; value: TValue }
+  | { success: false; error: string };
