+++ packages/twenty-server/src/engine/core-modules/file/file-upload/utils/to-batch-error-message.util.ts
@@ -1,0 +1,2 @@
+export const toBatchErrorMessage = (error: unknown): string =>
+  error instanceof Error ? error.message : String(error);
